refactor: complete refactor using separated packages dto.request and dto.response
