+++ back/healthapp/src/main/java/com/healthcare/domain/dto/request/PatientRequestDTO.java
@@ -1,7 +1,8 @@
-package com.healthcare.domain.dto;
+package com.healthcare.domain.dto.request;
 
 import java.time.LocalDate;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.healthcare.domain.model.enums.Gender;
 
 import jakarta.validation.constraints.NotBlank;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class PatientResponseDTO {
+public class PatientRequestDTO {
     @NotNull @NotBlank private String firstName;
     @NotNull @NotBlank private String lastName;
     @NotNull @NotBlank private String documentId;
@@ -23,4 +24,6 @@ public class PatientResponseDTO {
     @NotNull @NotBlank private String phone;
     @NotNull @NotBlank private String address;
     @NotNull @NotBlank private String emergencyContactInfo;
+    @JsonProperty("user")
+    @NotNull private UserRequestDTO user;
 }
