+++ back/healthapp/src/main/java/com/healthcare/domain/dto/request/UserRequestDTO.java
@@ -1,4 +1,4 @@
-package com.healthcare.domain.dto;
+package com.healthcare.domain.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
