+++ back/healthapp/src/main/java/com/healthcare/domain/model/entity/User.java
@@ -1,6 +1,6 @@
 package com.healthcare.domain.model.entity;
 
-import com.healthcare.domain.dto.UserRequestDTO;
+import com.healthcare.domain.dto.request.UserRequestDTO;
 import com.healthcare.domain.model.enums.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
